+++ src/main/java/com/jgvasconcelos/insurancebudget/domain/model/Car.java
@@ -3,6 +3,7 @@ package com.jgvasconcelos.insurancebudget.domain.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,7 +14,7 @@ public class Car {
     private String model;
     private String manufacturer;
     private Integer year;
-    private Float fipeValue;
+    private BigDecimal fipeValue;
     private List<Accident> accidents;
     private List<CarDriver> carDrivers;
     private LocalDateTime createdAt;
