+++ src/main/java/com/jgvasconcelos/insurancebudget/resources/repository/car/entity/CarEntity.java
@@ -16,6 +16,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.UuidGenerator;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -37,7 +38,7 @@ public class CarEntity {
     @Column(name = "year", nullable = false)
     private Integer year;
     @Column(name = "fipe_value", nullable = false)
-    private Float fipeValue;
+    private BigDecimal fipeValue;
     @OneToMany(mappedBy = "car", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     private List<AccidentEntity> accidents;
     @OneToMany(mappedBy = "car", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
